rework ported test
