+++ modules/apps/foundation/messaging/messaging-test/build.gradle
@@ -28,8 +28,6 @@ dependencies {
 
 	provided project(":apps:foundation:messaging:messaging-api")
 	provided project(":apps:foundation:messaging:messaging-impl")
-
-	provided group: 'org.osgi', name: 'org.osgi.core', version: '6.0.0'
 }
 
 task export_resolve(type: Resolve) {
